creating buy token order

+++ database/migrations/2018_12_11_223439_create_token_orders_table.php
@@ -13,22 +13,26 @@ class CreateTokenOrdersTable extends Migration
      */
     public function up()
     {
-        Schema::create('token_orders', function (Blueprint $table) {
-            $table->increments('order_id');
-            $table->integer("user_id");
-            $table->decimal("quantity");
-            $table->boolean("is_crypto");
-            $table->string("currency_symbol");
-            $table->string("status");
-            $table->timestamps();
-          //  $table->primary('order_id');
-            $table->foreign('user_id')
-            ->references('id')->on('users')
-            ->onDelete('cascade')->onUpdate('cascade');
+        if (Schema::hasTable('token_orders')==false) {
+            //
+            Schema::create('token_orders', function (Blueprint $table) {
+                $table->increments('order_id');
+                $table->integer("user_id");
+                $table->decimal("quantity");
+                $table->boolean("is_crypto");
+                $table->string("currency_symbol");
 
+                $table->string("status");
 
-        });
-        DB::statement('ALTER TABLE oken_orders AUTO_INCREMENT = 1003;');
+                $table->timestamps();
+              //  $table->primary('order_id');
+                $table->foreign('user_id')
+                ->references('id')->on('users')
+                ->onDelete('cascade')->onUpdate('cascade');
+            });
+            DB::statement('ALTER TABLE oken_orders AUTO_INCREMENT = 1003;');
+        }
+
     }
 
     /**
